Modularized Flask ATS project and improved frontend template
- Split monolithic app into modular structure with `app/`, `routes.py`, and `utils/`
- Added `__init__.py` and `create_app()` function for Flask factory pattern
- Moved core logic into separate utility modules:
  - `keyword_extraction.py` for extracting keywords from job description
  - `resume_matcher.py` for keyword matching logic with alias and semantic checks
  - `visualizer.py` for generating pie chart using Matplotlib
- Updated `routes.py` to handle the main route logic using Blueprint
- Created `templates/ats.html` with enhanced structure:
  - Integrated flash message system
  - Displayed resume vs JD keyword match results and pie chart
- Added layout improvements with placeholders for `style.css` and `favicon.ico`
- Verified working state with correct form handling and result rendering

These changes improve maintainability, readability, and prepare the app for further enhancements.
