+++ app/__init__.py
@@ -1,0 +1,15 @@
+import os
+from flask import Flask
+
+def create_app():
+    app = Flask(__name__)
+    app.secret_key = "your_secret_key"
+    app.config['UPLOAD_FOLDER'] = 'uploads'
+
+    if not os.path.exists(app.config['UPLOAD_FOLDER']):
+        os.makedirs(app.config['UPLOAD_FOLDER'])
+
+    from .routes import main
+    app.register_blueprint(main)
+
+    return app
