--- static/style.css
@@ -1,88 +1,0 @@
-body {
-    font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-    background-color: #f4f7f9;
-    margin: 0;
-    padding: 0;
-    color: #333;
-}
-
-.container {
-    width: 90%;
-    max-width: 800px;
-    margin: 40px auto;
-    background: #ffffff;
-    padding: 30px;
-    border-radius: 12px;
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.1);
-}
-
-h1, h2, h3 {
-    text-align: center;
-    color: #2c3e50;
-}
-
-.form-section {
-    display: flex;
-    flex-direction: column;
-    gap: 15px;
-    margin-bottom: 30px;
-}
-
-textarea, input[type="file"] {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid #ccd6dd;
-    border-radius: 6px;
-}
-
-button {
-    background-color: #4CAF50;
-    color: white;
-    padding: 12px;
-    border: none;
-    border-radius: 6px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-button:hover {
-    background-color: #45a049;
-}
-
-.flash-messages {
-    margin-bottom: 20px;
-}
-
-.flash {
-    padding: 12px;
-    border-radius: 5px;
-    margin-bottom: 10px;
-}
-
-.flash.success {
-    background-color: #dff0d8;
-    color: #3c763d;
-}
-
-.flash.warning {
-    background-color: #fcf8e3;
-    color: #8a6d3b;
-}
-
-.flash.danger {
-    background-color: #f2dede;
-    color: #a94442;
-}
-
-.results {
-    margin-top: 20px;
-    padding: 15px;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-}
-
-.chart {
-    display: block;
-    margin: 20px auto;
-    max-width: 300px;
-}
