--- templates/ats.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>ATS Score Analyzer</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
-</head>
-<body>
-    <div class="container">
-        <h1>ATS Score Analyzer</h1>
-        <form method="POST" enctype="multipart/form-data" class="form-section">
-            <label for="job_description">Job Description:</label>
-            <textarea id="job_description" name="job_description" rows="5" placeholder="Paste the job description here..." required></textarea>
-
-            <label for="resume">Upload Resume (PDF):</label>
-            <input type="file" id="resume" name="resume" accept=".pdf" required>
-
-            <button type="submit">Analyze</button>
-        </form>
-
-        {% with messages = get_flashed_messages(with_categories=true) %}
-            {% if messages %}
-                <div class="flash-messages">
-                    {% for category, message in messages %}
-                        <div class="flash {{ category }}">{{ message }}</div>
-                    {% endfor %}
-                </div>
-            {% endif %}
-        {% endwith %}
-
-        {% if chart_url %}
-        <div class="results">
-            <h2>Analysis Results</h2>
-            <p><strong>Match Percentage:</strong> {{ match_percentage|round(2) }}%</p>
-
-            <img src="data:image/png;base64,{{ chart_url }}" alt="Match Chart" class="chart">
-
-            <h3>Matched Keywords</h3>
-            {% if matched_terms %}
-                <ul>
-                    {% for keyword in matched_terms %}
-                        <li>{{ keyword }}</li>
-                    {% endfor %}
-                </ul>
-            {% else %}
-                <p>No keywords matched.</p>
-            {% endif %}
-
-            <h3>Missing Keywords</h3>
-            {% if unmatched_terms %}
-                <ul>
-                    {% for keyword in unmatched_terms %}
-                        <li>{{ keyword }}</li>
-                    {% endfor %}
-                </ul>
-            {% else %}
-                <p>All keywords matched.</p>
-            {% endif %}
-        </div>
-        {% endif %}
-    </div>
-</body>
-</html>
